Main page update
