+++ client/components/index.js
@@ -7,4 +7,5 @@ export {default as NavBar} from './NavBar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as Main} from './Main'
+export {default as BeerList} from './BeerList'
 export {default as SingleBeer} from './SingleBeer'
